Move stop command into the library
